Fixed problem with cast of SpiceInt to QVariant which occured only on decon
git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/trunk@2787 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/base/objs/Spice/Spice.cpp
@@ -268,7 +268,7 @@ namespace Isis {
           }
         }
 
-        QVariant result = frameCode;
+        QVariant result = (int)frameCode;
         storeValue("BODY_FRAME_CODE",0,SpiceIntType,result);
       }
       else {
